requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Explainability
grad-cam>=1.4.8

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-image>=0.21.0

# Data & Metrics
scikit-learn>=1.3.0
pandas>=2.0.0

# Configuration
PyYAML>=6.0

# Web App
Flask>=3.0.0
Werkzeug>=3.0.1

# Notebook
jupyter>=1.0.0
ipywidgets>=8.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
seaborn>=0.12.0

# Optional: Kaggle dataset download
# kaggle>=1.5.16
